# Conflicts:
#	designSystem/src/main/java/com/example/designsystem/theme/colors/AflamiColorScheme.kt
#	designSystem/src/main/java/com/example/designsystem/theme/colors/DarkColors.kt
#	designSystem/src/main/java/com/example/designsystem/theme/colors/LigthColors.kt
#	ui/src/main/java/com/example/ui/screens/movieDetails/MovieDetailsScreen.kt

+++ repository/src/main/java/com/example/repository/mapper/remote/CategoryRemoteMapper.kt
@@ -7,7 +7,7 @@ import com.example.repository.mapper.shared.EntityMapper
 class CategoryRemoteMapper : EntityMapper<RemoteCategoryDto, Category> {
     override fun toEntity(dto: RemoteCategoryDto): Category {
         return Category(
-            id = dto.id,
+            id = dto.id.toLong(),
             name = dto.name,
             imageUrl = ""
         )
